Allow some ansible libraries to be imported on Windows
The intention is to allow ansible-inventory-grapher to run on
windows. Windows-specific issues with other parts of the codebase
definitely still exist

Remove unix specific dependencies from ansible.utils to ansible.utils.unix
Removed unwanted pwd import from ansible.constants
Import display_function in ansible.utils only when required
Fixed broken path construction for loading host and group vars
Handling filenames that windows doesn't like (typically globs
for hosts in inventory - e.g. web*)
