+++ lib/ansible/constants.py
@@ -16,7 +16,6 @@
 # along with Ansible.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
-import pwd
 import sys
 import ConfigParser
 from string import ascii_letters, digits
@@ -88,7 +87,12 @@ def shell_expand_path(path):
 
 p = load_config_file()
 
-active_user   = pwd.getpwuid(os.geteuid())[0]
+try:
+    import pwd
+    active_user = pwd.getpwuid(os.geteuid())[0]
+except ImportError:
+    import getpass
+    active_user = getpass.getuser()
 
 # check all of these extensions when looking for yaml files for things like
 # group variables -- really anything we can load
